require setuptools >= 0.6c6 on cygwin, and >= 0.6a9 on other platforms
Earlier I tried 0.6a9 (which comes in .deb format on Dapper) and something
didn't work, so I raised it to 0.6c3.  However, I didn't make a note of what
failed, and so now I'm hoping that what failed was cygwin-specific.  Anyway,
everyone look out for setuptools compatibility problems on the your favorite
platform (and I'll check the buildslaves), and we'll raise the required version
as little as possible and only on the problematic platform.