fix test_vdrive (fixes #144)
It turns out that we actually have *two* files in our storage servers at the 
time that test_vdrive asserts things about the shares.  I suppose that 
test_vdrive happens to pass on all other operating systems because the 
filesystem happens to return the right share as the first one in a 
"listdir()".  The fix in this patch is slightly kludgey -- allow either share 
to pass -- but good enough.