confwiz: write accountname to a file
motivated simply by a desire to be able to identify 'noderoot' directories for
debugging and testing, the confwiz now writes an 'accountname' files based on 
what account was used when the node was configured.  this is not currently read
by or used by any code in the system, but helps identify directories from testing.